split annotation.ts types into separate files
