+++ src/PPSettings.interface.ts
